Refactor diff widgets
