+++ nbdime-web/src/diff/widget/common.ts
@@ -1,0 +1,11 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+'use strict';
+
+
+export const TWOWAY_DIFF_CLASS = 'jp-Diff-twoway';
+export const ADDED_DIFF_CLASS = 'jp-Diff-added';
+export const DELETED_DIFF_CLASS = 'jp-Diff-deleted';
+export const UNCHANGED_DIFF_CLASS = 'jp-Diff-unchanged';
+
+export const DIFF_CLASSES = ['jp-Diff-base', 'jp-Diff-remote'];
